measure 관련 util 수정
