+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './club';
 export * from './user';
 export * from './recommend';
+export * from './measuring';
